Fixed Enemy Linked List
Fixed the x variable which was unsigned and should have being signed, so it was causing the code to not check when the enemies where out of bounds through the left side of the screen hence never deleting items from the linked list

+++ Enemy.h
@@ -10,12 +10,12 @@ class Enemy{
 		Sprite enemySprite;
 		status_t status;
 		uint8_t live; // enemies staring life: 50 points of life for enemy 1, 100 points of life for enemy 2
-		uint8_t x;
+		int x;
 		uint8_t y;
 		uint8_t updatePosition;
 		uint8_t velocity; //we have different velocity depending on the enemy
 	public:
-		Enemy(uint8_t x, uint8_t y, uint8_t typeEnemy);//constructor, we specify the type of our enemy with it's initial live
+		Enemy(int x, uint8_t y, uint8_t typeEnemy);//constructor, we specify the type of our enemy with it's initial live
 		uint8_t getX();
 		uint8_t getY();
 		status_t getStatus();
@@ -23,7 +23,7 @@ class Enemy{
 		uint8_t getLive();
 		void reduceLive(uint8_t liveReduction);
 		void Draw(uint8_t hyper);
-		void UpdatePos(uint16_t x, uint16_t y);
+		void UpdatePos(int x, uint16_t y);
 		int8_t randomUpDownFn();
 	
 };
